\name{10SNPs.map}
\docType{data}
\alias{10SNPs.map}
\title{Plink sample PED data}
\description{
  This data set gives the genetic data in Plink format to be used for testing only.  This is to be used with 10SNPs.ped and Locations.txt.
}
%\usage{}
\format{A Plink PED format file.}
%\source{}
%\references{}
\keyword{datasets}
