\name{Locations}
\docType{data}
\alias{Locations}
\title{Locations of individuals in 10SNPs}
\description{
  This data set gives the locations of individuals in the Plink file 10SNPs to be used as a test data only.
}
\usage{Locations}
\format{A matrix containing names and locations.}
%\source{}
%\references{}
\keyword{datasets}
