\name{LocationsTrialDataSmall}
\docType{data}
\alias{LocationsTrialDataSmall}
\title{Locations of individuals in MicrosatTrialDataSmall}
\description{
  This data set gives the locations of individuals in the file MicrosatTrialDataSmall to be used as a test data only.  Space or tab delimited text file with the location information for the individuals.  The columns are LocationName, LocationNumber, Latitude, and Longitude.  Note that the first two columns must be in the same order as the MicrosatTrialDataSmall.
}
\usage{LocationsTrialDataSmall}
\format{A text file containing names and locations.}
%\source{}
%\references{}
\keyword{datasets}
