\name{MicrosatTrialDataSmall}
\docType{data}
\alias{MicrosatTrialDataSmall}
\title{Genotypes of individuals in located at LocationsTrialDataSmall}
\description{
  This data set gives the genotypes of individuals located in the file LocationsTrialDataSmall to be used as a test data only.  The columns here would be LocationName, LocationNumber, Locus1, Locus2, etc.  Each individual would take up 2 rows (one for each allele) with the same LocationName and LocationNumber.  The value under Locus would be the length of the allele of that individual.  Note that unknown individuals should have location number "-1".
}
\usage{MicrosatTrialDataSmall}
\format{A text file containing names and locations.}
%\source{}
%\references{}
\keyword{datasets}
